Refactor authentication and rework imports
