+++ Fridgify_Backend/models/backends/__init__.py
@@ -1,0 +1,2 @@
+from .user_authentication import UserAuthentication
+from .api_authentication import APIAuthentication
