+++ Fridgify_Backend/models/backends/api_authentication.py
@@ -8,13 +8,15 @@ from Fridgify_Backend.models import Accesstokens
 
 class APIAuthentication(authentication.BaseAuthentication):
     def authenticate(self, request):
-        token = request.headers["Authorization"]
+        req_token = request.headers["Authorization"]
         try:
             token = Accesstokens.objects.get(
-                accesstoken=token,
+                accesstoken=req_token,
                 provider="Fridgify-API",
                 valid_till__gte=time.time()
             )
+            token.user.is_authenticated = True
+            token.user.token_authentication = token.accesstoken
         except Accesstokens.DoesNotExist or Accesstokens.MultipleObjectsReturned:
             return exceptions.AuthenticationFailed
