+++ Fridgify_Backend/views/authentication/login.py
@@ -2,7 +2,7 @@ from rest_framework.decorators import api_view, authentication_classes, permissi
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 
-from Fridgify_Backend.models.backends.user_authentication import UserAuthentication
+from Fridgify_Backend.models.backends import UserAuthentication
 from Fridgify_Backend.utils import token_utils
 from Fridgify_Backend.utils.decorators import check_body
